Add files via upload
Task was to create code that could be malicious. From the user's perspective, this program asks for his/her date of birth to determine their Ghanaian name. However behind the scene, the program inserts random Twi text in all .py file in the directory this script is run , essentially altering the function of the user's py files.